Reorganization of the various experimental perf patches in more "thematic" patches, fixed a few things, added a bunch of library methods for querying part modules, models and renderers.
