+++ GameData/KSPCommunityFixes/Settings.cfg
@@ -420,9 +420,22 @@ KSP_COMMUNITY_FIXES
   // state synchronization and caching solar panels scaled space raycasts results.
   OptimizedModuleRaycasts = true
 
-  // General micro-optimization of FlightIntegrator and VesselPrecalculate, significantely increase
-  // framerate in large part count situations.
-  FlightPerf = true
+  // General micro-optimization of FlightIntegrator, VesselPrecalculate and FloatingOrigin, significantely 
+  // increase framerate in large part count situations.
+  FlightCoreSystemsPerf = true
+
+  // Faster lookup of other docking nodes
+  ModuleDockingNodeFindOtherNodesFaster = true
+
+  // Micro-optimizations of the CollisionEnhancer component, significant impact in high part count situations
+  CollisionEnhancerFastUpdate = true
+
+  // Micro-optimization of various flight scene auxiliary subsystems : temperature gauges, highlighter, 
+  // volume normalizer, strut position tracking...
+  AuxiliarySystemsFastUpdate = true
+
+  // Various small patches on hot loop methods
+  MinorPerfTweaks = true
 
   // ##########################
   // Modding
